+++ docs/CHANGES.md
@@ -1,3 +1,1 @@
-## v5.3.2
-- updated pkgmeta
-- remove extra \ from load order path mists toc
+- Update version numbers to include `v` prefix.